display user's existing rating for TV series and episodes

+++ data/remote_data_source/vod/src/main/java/com/sanaa/vod/media/tvShow/RemoteTvShowDataSourceImpl.kt
@@ -71,12 +71,12 @@ class RemoteTvShowDataSourceImpl(
         return apiService.fetchTvShowsGenres().genres.distinctBy { it.id }
     }
 
-    override suspend fun getTvShowRate(accountId: Int, sessionId: String): List<TvShowDto> =
+    override suspend fun getTvShowRate(accountId: Long, sessionId: String): List<TvShowDto> =
         wrapApiCall {
             apiService.fetchTvShowRate(accountId = accountId, sessionId = sessionId).results
         }
 
-    override suspend fun getEpisodesRate(accountId: Int, sessionId: String): List<EpisodeDto> =
+    override suspend fun getEpisodesRate(accountId: Long, sessionId: String): List<EpisodeDto> =
         wrapApiCall {
             apiService.fetchEpisodesRate(sessionId = sessionId, accountId = accountId).results
         }
